controller/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
httpx==0.26.0
google-generativeai==0.4.0
pyyaml==6.0.1
